+++ WebClient/MiniFile/MiniFile.ts
@@ -1,0 +1,22 @@
+import Vue from 'vue';
+import Component from 'vue-class-component';
+import { Inject, Watch, Prop } from 'vue-property-decorator';
+import { Object } from 'common/Files';
+
+@Component({
+    components: {
+    },
+})
+export default class MiniFile extends Vue {
+
+    @Prop() file: Object;
+    @Prop({ default: false }) selected: boolean;
+
+    constructor() {
+        super();
+    }
+
+    click() {
+        this.$emit('click');
+    }
+};