+++ WebClient/Welcome/Welcome.vue
@@ -10,8 +10,8 @@
         <md-card-content>
           <div>
             <md-field>
-              <label for="name">Email</label>
-              <md-input name="email" id="email" autocomplete="email" v-model="email"/>
+              <label for="name">Name</label>
+              <md-input name="name" id="name" autocomplete="name" v-model="email"/>
             </md-field>
           </div>
           <div class="continue-as-guest-section">