+++ WebClient/game-engine/Input.ts
@@ -1,3 +1,4 @@
+import Vue from 'vue';
 import { ArgEvent } from '../../common/Events';
 import { Vector2, Vector3 } from 'three';
 import { find, findIndex, some } from 'lodash';
@@ -14,9 +15,14 @@ export class Input {
     private _mouseData: MouseData;
     private _touchData: TouchData[];
     private _wheelData: WheelData;
+    private _targetData: TargetData;
 
     private _gameView: GameView;
     private _inputType: InputType = InputType.Undefined;
+
+    // Keep track of the touch data for finger index 0 at all times.
+    // This gives better support to the getMouse* functions while using touch.
+    private _lastPrimaryTouchData: TouchData;
 
     /**
      * @returns 'mouse' or 'touch'
@@ -82,8 +88,21 @@ export class Input {
             pagePos: new Vector2(0, 0),
             clientPos: new Vector2(0, 0)
         };
+        this._targetData = {
+            inputDown: null,
+            inputUp: null,
+            inputOver: null
+        };
         this._touchData = [];
         this._wheelData = new WheelData();
+        this._lastPrimaryTouchData = { 
+            fingerIndex: 0,
+            identifier: 0,
+            clientPos: new Vector2(0,0),
+            pagePos: new Vector2(0,0),
+            screenPos: new Vector2(0,0),
+            state: new InputState()
+        };
 
         this._handleMouseDown = this._handleMouseDown.bind(this);
         this._handleMouseMove = this._handleMouseMove.bind(this);
@@ -95,7 +114,7 @@ export class Input {
         this._handleTouchCancel = this._handleTouchCancel.bind(this);
         this._handleContextMenu = this._handleContextMenu.bind(this);
 
-        let element = this._gameView.gameView;
+        let element = document.body;
         element.addEventListener('mousedown', this._handleMouseDown);
         element.addEventListener('mousemove', this._handleMouseMove);
         element.addEventListener('mouseup', this._handleMouseUp);
@@ -104,14 +123,16 @@ export class Input {
         element.addEventListener('touchmove', this._handleTouchMove);
         element.addEventListener('touchend', this._handleTouchEnd);
         element.addEventListener('touchcancel', this._handleTouchCancel);
-        element.addEventListener('contextmenu', this._handleContextMenu);
+        
+        // Context menu is only important on the game view
+        this._gameView.gameView.addEventListener('contextmenu', this._handleContextMenu);
 
     }
 
     public dispose() {
         console.log("[Input] dispose");
 
-        let element = this._gameView.gameView;
+        let element = document.body;
         element.removeEventListener('mousedown', this._handleMouseDown);
         element.removeEventListener('mousemove', this._handleMouseMove);
         element.removeEventListener('mouseup', this._handleMouseUp);
@@ -120,9 +141,61 @@ export class Input {
         element.removeEventListener('touchmove', this._handleTouchMove);
         element.removeEventListener('touchend', this._handleTouchEnd);
         element.removeEventListener('touchcancel', this._handleTouchCancel);
-        element.removeEventListener('contextmenu', this._handleContextMenu);
+        
+        // Context menu is only important on the game view
+        this._gameView.gameView.removeEventListener('contextmenu', this._handleContextMenu);
 
         this._gameView = null;
+    }
+
+    /**
+     * Determines if the mouse down event happened directly over the given element.
+     * @param element The element to test.
+     */
+    public isMouseButtonDownOn(element: HTMLElement): boolean {
+        const downElement = this._targetData.inputDown;
+        return Input.isElementContainedByOrEqual(downElement, element);
+    }
+
+    /**
+     * Determines if the mouse is currently focusing the given html element.
+     * @param element 
+     */
+    public isMouseFocusing(element: HTMLElement): boolean {
+        const overElement = this._targetData.inputOver;
+        return Input.isElementContainedByOrEqual(overElement, element);
+    }
+
+    /**
+     * Gets the closest vue component accociated with this HTML element.
+     * @param element The html element.
+     */
+    public static getVueParent(element: HTMLElement): Vue {
+        const e = <any>element;
+        if (!e) {
+            return null;
+        }
+        if (e.__vue__) {
+            return <Vue>e.__vue__;
+        }
+        return Input.getVueParent(element.parentElement);
+    }
+
+    /**
+     * Determines if the given HTML element is contained by the given container element.
+     * @param element The HTML element.
+     * @param container The container.
+     */
+    public static isElementContainedByOrEqual(element: HTMLElement, container: HTMLElement): boolean {
+        if (element === container) {
+            return true;
+        } else {
+            if (!element) {
+                return false;
+            } else {
+                return this.isElementContainedByOrEqual(element.parentElement, container);
+            }
+        }
     }
 
     /**
@@ -137,7 +210,7 @@ export class Input {
             }
         } else if (this._inputType == InputType.Touch) {
             if (buttonId == MouseButtonId.Left) {
-                return this.getTouchDown(0);
+                return this._lastPrimaryTouchData.state.isDownOnFrame(this._gameView.time.frameCount);
             } else {
                 // TODO: Support right button with touch?
             }
@@ -167,7 +240,7 @@ export class Input {
             }
         } else if (this._inputType == InputType.Touch) {
             if (buttonId == MouseButtonId.Left) {
-                this.getTouchUp(0);
+                return this._lastPrimaryTouchData.state.isUpOnFrame(this._gameView.time.frameCount);
             } else {
                 // TODO: Support right button with touch?
             }
@@ -197,7 +270,7 @@ export class Input {
             }
         } else if (this._inputType == InputType.Touch) {
             if (buttonId == MouseButtonId.Left) {
-                return this.getTouchHeld(0);
+                return this._lastPrimaryTouchData.state.isHeldOnFrame(this._gameView.time.frameCount);
             } else {
                 // TODO: Support right button with touch?
             }
@@ -240,7 +313,7 @@ export class Input {
         if (this._inputType == InputType.Mouse) {
             return this._mouseData.screenPos;
         } else if (this._inputType == InputType.Touch) {
-            return this.getTouchScreenPos(0);
+            return this._lastPrimaryTouchData.screenPos;
         }
 
         return null;
@@ -254,7 +327,7 @@ export class Input {
         if (this._inputType == InputType.Mouse) {
             return this._mouseData.pagePos;
         } else if (this._inputType == InputType.Touch) {
-            return this.getTouchPagePos(0);
+            return this._lastPrimaryTouchData.pagePos;
         }
 
         return null;
@@ -268,7 +341,7 @@ export class Input {
         if (this._inputType == InputType.Mouse) {
             return this._mouseData.clientPos;
         } else if (this._inputType == InputType.Touch) {
-            return this.getTouchClientPos(0);
+            return this._lastPrimaryTouchData.clientPos;
         }
 
         return null;
@@ -336,6 +409,16 @@ export class Input {
         return null;
     }
 
+    /**
+     * Gets the information about what HTML elements are currently being targeted.
+     * Note that this only stores information about the last targeted elements.
+     * As such, it should only be used to tell whether touch/mouse events
+     * should be used or not.
+     */
+    public getTargetData(): TargetData {
+        return this._targetData;
+    }
+
     public update() {
 
         this._cullTouchData();
@@ -383,6 +466,15 @@ export class Input {
         }
     }
 
+    private _copyToPrimaryTouchData(data: TouchData) {
+        this._lastPrimaryTouchData.fingerIndex = data.fingerIndex;
+        this._lastPrimaryTouchData.identifier = data.identifier;
+        this._lastPrimaryTouchData.clientPos = data.clientPos.clone();
+        this._lastPrimaryTouchData.pagePos = data.pagePos.clone();
+        this._lastPrimaryTouchData.screenPos = data.screenPos.clone();
+        this._lastPrimaryTouchData.state = data.state.clone();
+    }
+
     /**
      * Returns the matching MouseButtonData object for the provided mouse button number.
      */
@@ -417,6 +509,7 @@ export class Input {
                 console.log("mouse button " + event.button + " down. fireInputOnFrame: " + this._gameView.time.frameCount);
             }
 
+            this._targetData.inputDown = <HTMLElement>event.target;
             this._mouseData.clientPos = new Vector2(event.clientX, event.clientY);
             this._mouseData.pagePos = new Vector2(event.pageX, event.pageY);
             this._mouseData.screenPos = this._calculateScreenPos(event.pageX, event.pageY);
@@ -435,6 +528,7 @@ export class Input {
                 console.log("mouse button " + event.button + " up. fireInputOnFrame: " + this._gameView.time.frameCount);
             }
 
+            this._targetData.inputUp = <HTMLElement>event.target;
             this._mouseData.clientPos = new Vector2(event.clientX, event.pageY);
             this._mouseData.pagePos = new Vector2(event.pageX, event.pageY);
             this._mouseData.screenPos = this._calculateScreenPos(event.pageX, event.pageY);
@@ -448,6 +542,7 @@ export class Input {
         this._mouseData.clientPos = new Vector2(event.clientX, event.clientY);
         this._mouseData.pagePos = new Vector2(event.pageX, event.pageY);
         this._mouseData.screenPos = this._calculateScreenPos(event.pageX, event.pageY);
+        this._targetData.inputOver = <HTMLElement>event.target;
 
         if (this.debugLevel >= 2) {
             console.log("mouse move:");
@@ -458,7 +553,9 @@ export class Input {
     }
 
     private _handleWheel(event: WheelEvent) {
-        event.preventDefault();
+        if (this.isMouseFocusing(this._gameView.gameView)) {
+            event.preventDefault();
+        }
 
         let wheelFrame: WheelFrame = {
             moveFrame: this._gameView.time.frameCount,
@@ -500,10 +597,15 @@ export class Input {
             // Set the down frame on the new touch data.
             data.state.setDownFrame(this._gameView.time.frameCount);
 
+            if (data.fingerIndex === 0) {
+                this._copyToPrimaryTouchData(data);
+            }
+
             if (this.debugLevel >= 1) {
                 console.log("touch finger " + data.fingerIndex + " start. fireInputOnFrame: " + this._gameView.time.frameCount);
             }
 
+            this._targetData.inputDown = <HTMLElement>touch.target;
             this._touchData.push(data);
         }
     }
@@ -526,6 +628,10 @@ export class Input {
             existingTouch.pagePos = new Vector2(touch.pageX, touch.pageY);
             existingTouch.screenPos = this._calculateScreenPos(touch.pageX, touch.pageY);
 
+            if (existingTouch.fingerIndex === 0) {
+                this._copyToPrimaryTouchData(existingTouch);
+            }
+
             if (this.debugLevel >= 2) {
                 console.log("touch move:");
                 console.log("  identifier: " + existingTouch.identifier);
@@ -547,11 +653,17 @@ export class Input {
         for (let i = 0; i < changed.length; i++) {
             let touch = changed.item(i);
 
+            this._targetData.inputUp = <HTMLElement>touch.target;
+
             let existingTouch = find(this._touchData, (d) => { return d.identifier === touch.identifier; });
             existingTouch.state.setUpFrame(this._gameView.time.frameCount);
             existingTouch.clientPos = new Vector2(touch.clientX, touch.clientY);
             existingTouch.pagePos = new Vector2(touch.pageX, touch.pageY);
             existingTouch.screenPos = this._calculateScreenPos(touch.pageX, touch.pageY);
+            
+            if (existingTouch.fingerIndex === 0) {
+                this._copyToPrimaryTouchData(existingTouch);
+            }
 
             if (this.debugLevel >= 1) {
                 console.log("touch finger " + existingTouch.fingerIndex + " end. fireInputOnFrame: " + this._gameView.time.frameCount);
@@ -574,6 +686,10 @@ export class Input {
             existingTouch.clientPos = new Vector2(touch.clientX, touch.clientY);
             existingTouch.pagePos = new Vector2(touch.pageX, touch.pageY);
             existingTouch.screenPos = this._calculateScreenPos(touch.pageX, touch.pageY);
+
+            if (existingTouch.fingerIndex === 0) {
+                this._copyToPrimaryTouchData(existingTouch);
+            }
 
             if (this.debugLevel >= 1) {
                 console.log("touch finger " + existingTouch.fingerIndex + " canceled. fireInputOnFrame: " + this._gameView.time.frameCount);
@@ -662,6 +778,17 @@ export class InputState {
     isUpOnFrame(frame: number): boolean {
         return frame === this._upFrame;
     }
+
+    /**
+     * Returns a new InputState with the same values as this one.
+     */
+    clone(): InputState {
+        let clone = new InputState();
+        clone._downFrame = this._downFrame;
+        clone._upFrame = this._upFrame;
+
+        return clone;
+    }
 }
 
 interface WheelFrame {
@@ -683,6 +810,15 @@ interface WheelFrame {
      * This boolean indicates whether the ctrl key was detected with the WheelEvent.
      */
     ctrl: boolean;
+}
+
+/**
+ * Data about the HTML element that was targeted by a click.
+ */
+interface TargetData {
+    inputDown: HTMLElement;
+    inputUp: HTMLElement;
+    inputOver: HTMLElement;
 }
 
 interface TouchData {